Resolving some bugs in the canvas, and put the canvas in one component
